Fix setup order to install Ingress at the end

+++ src/helm/setup/_install_jizt.sh
@@ -4,33 +4,36 @@ JIZT_RELEASE_NAME="jizt"
 JIZT_NAMESPACE="default"  # must match JIZT_NAMESPACE in _instal_strimzi_operator,sh
 JIZT_VERSION="0.0.1"
 JIZT_CHART_PATH="../jizt"  # don't add trailing slash
+POST_INSTALL=( '*deployment*' '*ingress*' )  # components to be installed the last
 
 # If user interrupts installation, make sure we leave
 # everything consistent.
 trap tidy_up SIGINT  
 tidy_up() {
-    unhide_deployments
+    unhide_components
     exit 1
 }
 
 # Function definitions
-hide_deployments() {
+hide_components() {
     cd "$JIZT_CHART_PATH/templates"
-    for file in *-deployment.yaml; do
+    for file in ${POST_INSTALL[@]}; do
         if [[ -f $file  ]]; then
-            mv $file .$file  # hide files
+            mv "$file" ".$file"  # hide files
         fi
     done
     cd - > /dev/null  # don't print directory
 }
 
-unhide_deployments() {
+unhide_components() {
     cd "$JIZT_CHART_PATH/templates"
-    for file in .*-deployment.yaml; do
-        if [[ -f $file  ]]; then
+    shopt -s dotglob  # enable globbing for hidden files
+    for file in ${POST_INSTALL[@]}; do
+        if [[ -f $file && ${file::1} == '.'  ]]; then
             mv $file "${file:1}"  # unhide files
         fi;
     done
+    shopt -u dotglob  # disable globbing for hidden files
     cd - > /dev/null  # don't print directory
 }
 
@@ -42,9 +45,9 @@ kubectl create namespace $JIZT_NAMESPACE 2> /dev/null
 echo -e "\n>>> Installing Strimzi/Kafka related components in namespace '$JIZT_NAMESPACE'..."
 
 # Helm "--wait" flag does not seem to give much of a sh*t for Strimzi stuff so
-# we hide microservices deployments so Helm doesn't install them yet.
-hide_deployments
-sleep 8
+# we hide microservices deployments and Ingress so Helm doesn't install them yet.
+hide_components
+
 cat << EOF | helm install $JIZT_RELEASE_NAME $JIZT_CHART_PATH \
 --wait \
 --namespace $JIZT_NAMESPACE \
@@ -66,9 +69,9 @@ done
 
 echo -e "\n>>> Installing JIZT microservices in namespace '$JIZT_NAMESPACE'..."
 
-# Unhide microservices deployments to install them now
-# that everything is ready.
-unhide_deployments
+# Unhide microservices deployments and Ingress to install them now that everything is ready
+unhide_components
+
 cat << EOF | helm upgrade $JIZT_RELEASE_NAME $JIZT_CHART_PATH \
 --wait \
 --namespace $JIZT_NAMESPACE \
